Change resources files handling
The resource files that have to be read and written are now placed in directories next to the jar file; they are created by the AS at runtime, and deleted before creating them again in the next run.
Files that have to be read only are left within the jar.

This solution lets the code run without errors when the AS is executed both from the IDE and from the command line launching the jar.